chore(sdk): move viz from top level
